fix for small int
adjusted Small int for SQL Server plus considering making the generic data not a Universal for all the databases due to type quirks

+++ src/internals/data_structures/database_metadata/table_data.rs
@@ -5,7 +5,7 @@ use tiberius::time::chrono::{self, DateTime, NaiveDateTime};
 #[derive(Debug)]
 pub enum GenericData {
     Text(Option<String>),
-    SmallInt(Option<i8>),
+    SmallInt(Option<i16>),
     Int(Option<i32>),
     Float(Option<f64>),
     Bool(Option<bool>),
